Providing after_build and after_create callbacks for factories.
This allows you to define callbacks that will be passed the instance that has just been built or created by the factory, allowing you to manipulate the instance beyond the realms of simply setting attributes and associations.  The intention is to allow for stubbing or mocking of methods that would otherwise require a complex graph of attributes and associations.  There may be other uses too, such as manipulating the object outside the realm of it's associations.